+++ @xen-orchestra/web-core/lib/icons/action-icons.ts
@@ -1,3 +1,4 @@
+import { slash } from '@core/icons/custom-icons.ts'
 import { defineIconPack, type IconSingleConfig } from '@core/packages/icon'
 import { faCopy, faSquare as checkboxEmpty, type IconDefinition } from '@fortawesome/free-regular-svg-icons'
 import {
@@ -24,6 +25,7 @@ import {
   faLink,
   faLinkSlash,
   faMinus,
+  faPlug,
   faPowerOff,
   faRefresh,
   faRoute,
@@ -132,8 +134,10 @@ export const actionIcons = defineIconPack({
   fill: constructIcon(faFillDrip),
   duplicate: constructIcon(faClone),
   copy: constructIcon(faCopy),
-  connect: constructIcon(faLink),
-  disconnect: constructIcon(faLinkSlash),
+  attach: constructIcon(faLink),
+  detach: constructIcon(faLinkSlash),
+  connect: constructIcon(faPlug),
+  disconnect: [constructIcon(faPlug), { icon: slash }],
   forget: constructIcon(faEraser),
   delete: { icon: faTrash, color: 'var(--color-danger-txt-primary)' },
   'more-actions': constructIcon(faEllipsis),