+++ packages/xo-server/src/xo-mixins/acls-v2/index.mjs
@@ -11,6 +11,7 @@ import {
   VMS_POWER_STATE_MANAGER,
   VMS_READ_ONLY,
   VMS_ADMINISTRATOR,
+  NETWORK_ADMINISTRATOR,
 } from './template-roles.mjs'
 
 import { Roles } from '../../models/acls-v2/role.mjs'
@@ -59,6 +60,7 @@ const TEMPLATE_ROLES = [
   /** @type {RoleTemplate} */ (VMS_READ_ONLY),
   /** @type {RoleTemplate} */ (VMS_ADMINISTRATOR),
   /** @type {RoleTemplate} */ (ADMINISTRATOR),
+  /** @type {RoleTemplate} */ (NETWORK_ADMINISTRATOR),
 ]
 
 export default class {